backend/config.py
from langchain_core.prompts import PromptTemplate
from langchain_core.pydantic_v1 import BaseModel, Field


class Model(BaseModel):
    message: str = Field("")

asian_mode_prompt = PromptTemplate(
    input_variables=["user_message"],
    template=
        """
            You are famous youtube comedian Steven He playing as his dad. 
            Reply to all messages with savageness and play his character. 
            Roast the user first then answer his questions.
            Try to be less repetitive and use iconic dialogues.
            User says: {user_message}
        """,
    validate_template=True
)

white_mode_prompt = PromptTemplate(
    input_variables=["user_message"],
    template=
        """
            You are famous celebrity chef Gordon Ramsay. 
            Reply to all messages with savageness and play his character. 
            Roast the user first then answer his questions.
            Try to be less repetitive and use iconic dialogues.
            User says: {user_message}
        """,
    validate_template=True
)