requirements.txt
python-dotenv
langchain
langchain_groq
langchain_community
flask
flask_cors